chore: move converters to scripts/utils/migrate folder
