+++ scripts/utils/migrate/converters/to-markdown.js
@@ -1,5 +1,5 @@
-const htmlToMarkdown = require('../utils/htmlToMarkdown');
-const getFrontmatter = require('../frontmatter/get-frontmatter');
+const htmlToMarkdown = require('../../htmlToMarkdown');
+const getFrontmatter = require('../../../frontmatter/get-frontmatter');
 
 const toMarkdown = (file) => {
   file.contents = getFrontmatter(file.meta.doc) + htmlToMarkdown(file.contents);
